feature-policy/ch-dpr-disabled-by-feature-policy.tentative.https.sub.html

<!DOCTYPE html>
<body>
  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>
  <script src=/feature-policy/resources/featurepolicy.js></script>
  <script>
  'use strict';
  var same_origin_src = '/feature-policy/resources/feature-policy-ch-dpr-cross-origin-resource.sub.html';
  var cross_origin_src = 'https://{{domains[www]}}:{{ports[https][0]}}' +
    same_origin_src;
  var header = 'Feature-Policy header {"ch-dpr" : ["*"]}';

  promise_test(() => {
    return fetch('https://{{domains[www]}}:{{ports[https][0]}}/client-hints/echo_client_hints_received.py').then(r => {
      window.h = r.headers;
      assert_equals(r.status, 200);
      assert_false(r.headers.has('dpr-received'));
    });
  }, header + ' disallows the top-level document.');

  async_test(t => {
    test_feature_availability('The ch-dpr feature policy', t, same_origin_src,
        expect_feature_unavailable_default);
  }, header + ' disallows same-origin iframes.');

  async_test(t => {
    test_feature_availability('The ch-dpr feature policy', t, cross_origin_src,
        expect_feature_unavailable_default);
  }, header + ' disallows cross-origin iframes.');

  </script>
</body>
